分两个文件，手动move constructor，单一触发多特效
触发多特效支持双手互搏计数
切记如果需要move vector，不能 return { readXX() }; 作list initialization
修复临时状态没清空bug
加入几个特效
